Refactor p2p key module (#115)
* remove identity.go in p2p
* rename peerKey.go to p2p_key.go
* rename functions in p2p_key.go
* add test cases
* Add license in documentation
* Fix p2p test case for fast sync
